Conflicts:
	Banka/ant.properties
	CentralnaBanka/build.properties
	CentralnaBanka/src/session/dao/Mt10xDaoBean.java
	CentralnaBanka/src/session/dao/Mt9xyDaoBean.java
	CentralnaBanka/src/session/dao/Mt9xyDaoLocal.java

+++ CentralnaBanka/src/entity/Banka.java
@@ -88,7 +88,7 @@ public class Banka{
 		this.naziv = naziv;
 	}
 
-	public double getSifra() {
+	public int getSifra() {
 		return sifra;
 	}
